minor args restrucuring
